Implemented StandardNetworkableEvent System. Allows for Synchronisation of classes that Implement StandardNetworkableEvent.
A base class NetworkableEvent has been added to allow the EventHandler that catches NetworkableEvents (to be implemented) to know which IMessage he has to send. The derived StandardNetworkableEvent has the StandardNetworkableEventSyncMessage as its IMessage.

It is up to the User to register his additional IMessages to the EventNetworkManagers Channel. The StandardNetworkableEventSyncMessage is preregistered by SmithsCore regardless if it is used.

The EventHandler that will be responsible for sending the message will be a lowest priority one, meaning that it will fire last is it is not Canceled before hand.