Continue the implementation of the ITabManager and CoreTab.
The whole Idea behind the Tab system is a wrapper mechanic that replaces the current direct registration of the components to the UI it self. Instead if no Tabs are registered a DummyTab will be created and the implementing UI will add the components to that tab (as there is only one tab then, no information for the selector is needed) which is then rendered as one.