Added IItemStorage to replace IInventory.
Added a replacement for IInventory, as it has a lot of methods that are not used or SmithsCore has replaced with different ones.
Fixed the PlayerManager not having all methods when it runs in the InternalServer.
Fixed handleKeyTyped not reporting back if the KeyType was consumed by any components.
Fixed the GuiInputEvent not being processed on the server side, if type was not TABCHANGED.